third_party/WebKit/LayoutTests/fast/dom/HTMLTableElement/exceptions-expected.txt

This test should trigger exceptions on HTMLTableElement, and verify that the messages are reasonably helpful.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS t.insertRow(-2) threw exception IndexSizeError: Failed to execute 'insertRow' on 'HTMLTableElement': The index provided (-2) is less than -1..
PASS t.insertRow(1) threw exception IndexSizeError: Failed to execute 'insertRow' on 'HTMLTableElement': The index provided (1) is greater than the number of rows in the table (0)..
PASS t.deleteRow(-2) threw exception IndexSizeError: Failed to execute 'deleteRow' on 'HTMLTableElement': The index provided (-2) is less than -1..
PASS t.deleteRow(2) threw exception IndexSizeError: Failed to execute 'deleteRow' on 'HTMLTableElement': The index provided (2) is greater than the number of rows in the table (1)..
PASS t.deleteRow() threw exception TypeError: Failed to execute 'deleteRow' on 'HTMLTableElement': 1 argument required, but only 0 present..
PASS successfullyParsed is true

TEST COMPLETE
